pyproject.toml
[project]
name = "ease-hs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.37.23",
    "monsterui>=1.0.11",
    "python-dotenv>=1.1.0",
    "python-fasthtml>=0.12.5",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
